Allow player to add token in a different location
This is done only if the location that the player chose already has a
token. Previously, if a location was occupied, a message stating such
was displayed onto the console and the game directly jumped onto the
next round with a different active player and the previous player would
lose his chance of  adding a token. This behavior has hence been
rectified and the player is prompted for a location till he/she enters
a location that is not already occupied by a token.