Change names given to the players to generic names
Previously, we had hardcoded the player names to unique values. Now
these are changed to generic values "Player X" & "Player O" since the
users are now allowed to enter the player names and also since each
player having two unique names is not exactly ideal. Even though the
generic names will be used in the code,the user entered names will be
displayed on the screen.